+++ contributions/api/models/talents/talent.py
@@ -18,6 +18,7 @@ class Talent():
     def __init__(self, injson):
         self.id = None
         self.name = None
+        self.icon = None
         self.shortDescription = None
         self.longDescription = None
         self.requiredCapabilities = None
@@ -36,6 +37,12 @@ class Talent():
 
     def get_id(self):
         return self.id
+
+    def set_icon(self, icon):
+        self.icon = icon
+
+    def get_icon(self):
+        return self.icon
 
     def set_name(self, name):
         self.name = name